pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "foldify"
version = "0.1.0"
description = "AI-powered local file organizer — sort your files intelligently using rules and local LLMs"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Youness Berkia", email = "berkiayouness@gmail.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Filesystems",
    "Topic :: Utilities",
]
keywords = ["foldify", "file", "organizer", "ai", "ollama", "automation", "cli", "local-ai"]
requires-python = ">=3.10"
dependencies = [
    "click>=8.0.0",
    "pyyaml>=6.0",
    "ollama>=0.1.0",
    "pypdf>=3.0.0",
    "python-docx>=0.8.11",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
    "build>=1.2.0",
    "types-PyYAML>=6.0.12",
]

[project.scripts]
foldify = "foldify.cli:main"

[project.urls]
Homepage = "https://github.com/YounessBerkia/foldify"
Repository = "https://github.com/YounessBerkia/foldify"
Issues = "https://github.com/YounessBerkia/foldify/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
foldify = ["py.typed"]

[tool.black]
line-length = 88
target-version = ['py310']

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = "yaml"
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=foldify --cov-report=term-missing"
